refactor: add addLike and removeLike methods

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -33,7 +33,7 @@ public class UserService {
      */
 
     private final UserStorage userStorage;
-    private int userId = 0;
+    private long userId = 0;
 
     @Autowired
     public UserService(UserStorage userStorage) {
